% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfExplorer2-package.R
\docType{package}
\name{adfExplorer2-package}
\alias{adfExplorer2}
\alias{adfExplorer2-package}
\title{adfExplorer2: Access and Manipulate Amiga Disk Files}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Amiga Disk Files (ADF) are virtual representations of 3.5 inch floppy disks for the Commodore Amiga. Most disk drives from other systems (including modern drives) are not able to read these disks. The AdfExplorer2 package enables you to establish R connections to files on such virtual DOS-formatted disks, which can be use to read from and write to those files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pepijn-devries.github.io/adfExplorer2/}
}

}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})

}
\keyword{internal}
