% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{adf_dir_list}
\alias{adf_dir_list}
\alias{adf_dir_list.adf_device}
\alias{adf_dir_list.adf_device.character}
\alias{adf_dir_list.adf_device.virtual_path}
\title{List entries in a directory of a virtual ADF device}
\usage{
adf_dir_list(dev, virtual_path, recursive = FALSE, ...)

\method{adf_dir_list}{adf_device}(dev, virtual_path, recursive = FALSE, ...)

\method{adf_dir_list.adf_device}{character}(dev, virtual_path, recursive = FALSE, ...)

\method{adf_dir_list.adf_device}{virtual_path}(dev, virtual_path, recursive = FALSE, ...)
}
\arguments{
\item{dev}{The virtual adf device for which information needs to be obtained.
It should be of class \code{adf_device} which can be created with \code{\link[=create_adf_device]{create_adf_device()}}
or \code{\link[=connect_adf]{connect_adf()}}.}

\item{virtual_path}{The virtual path for which you wish to obtain a list
of entries (see also \code{vignette("virtual_paths")}). When missing,
entries for the current directory (\code{\link[=adf_directory]{adf_directory()}}) are returned.}

\item{recursive}{A \code{logical} value. When set to \code{TRUE}, the function is
called recursively for all subdirectories in \code{virtual_path}.}

\item{...}{Ignored}
}
\value{
A vector of \code{virtual_path} class objects
}
\description{
Get an overview of all entries (files and directories) in a specific
directory.
}
\examples{
## First setup a connection to a virtual device
adz_file <- system.file("example.adz", package = "adfExplorer2")
my_device <- connect_adf(adz_file)

## List all entries in the disk's root:
adf_dir_list(my_device)
## List all entries on the disk:
adf_dir_list(my_device, recursive = TRUE)

close(my_device)
}
\author{
Pepijn de Vries
}
